# v0.5.0 -- internal clarity