--- a/query/parser.go
+++ b/query/parser.go
@@ -1,6 +1,7 @@
 package query
 
 import (
+	"fmt"
 	"io"
 	"sort"