- Fixes a ton of issues with correctness in the query processor
- Reimplements the SPARQL parser to more closely follow the actual grammar
- faster!

Benchmarks:
```
benchmark                                               old ns/op     new ns/op     delta
BenchmarkQueryPerformance1/SimpleSubjectVarTriple-4     1357831       606024        -55.37%
BenchmarkQueryPerformance1/LongerQuery1-4               14132902      4573599       -67.64%
BenchmarkQueryPerformance1/LooseQuery-4                 824859736     8762289       -98.94%
BenchmarkQueryPerformance1/LocQuery-4                   17617665      12537599      -28.84%
BenchmarkQueryPerformance1/RoomEnum-4                   1203879       708075        -41.18%
BenchmarkQueryPerformance1/AHUFeed0-4                   1135832       1116922       -1.66%
BenchmarkQueryPerformance1/AHUFeed1-4                   2938893       1969657       -32.98%
BenchmarkQueryPerformance1/AHUFeed2-4                   3629817       2393530       -34.06%
BenchmarkQueryPerformance1/AHUFeed1Reverse-4            4142404       3324894       -19.74%
BenchmarkQueryPerformance1/SensorSubclass-4             1404690       1302003       -7.31%
BenchmarkQueryPerformance1/VAVExplore-4                 5474127       6173419       +12.77%

benchmark                                               old allocs     new allocs     delta
BenchmarkQueryPerformance1/SimpleSubjectVarTriple-4     5442           2755           -49.38%
BenchmarkQueryPerformance1/LongerQuery1-4               36284          29076          -19.87%
BenchmarkQueryPerformance1/LooseQuery-4                 1219817        22222          -98.18%
BenchmarkQueryPerformance1/LocQuery-4                   49391          78801          +59.55%
BenchmarkQueryPerformance1/RoomEnum-4                   4471           2259           -49.47%
BenchmarkQueryPerformance1/AHUFeed0-4                   3032           3456           +13.98%
BenchmarkQueryPerformance1/AHUFeed1-4                   11609          6544           -43.63%
BenchmarkQueryPerformance1/AHUFeed2-4                   12957          6569           -49.30%
BenchmarkQueryPerformance1/AHUFeed1Reverse-4            12707          6551           -48.45%
BenchmarkQueryPerformance1/SensorSubclass-4             4634           3030           -34.61%
BenchmarkQueryPerformance1/VAVExplore-4                 25360          34867          +37.49%

benchmark                                               old bytes     new bytes     delta
BenchmarkQueryPerformance1/SimpleSubjectVarTriple-4     1448817       228484        -84.23%
BenchmarkQueryPerformance1/LongerQuery1-4               2895743       1174642       -59.44%
BenchmarkQueryPerformance1/LooseQuery-4                 263754992     1880167       -99.29%
BenchmarkQueryPerformance1/LocQuery-4                   3725373       2888308       -22.47%
BenchmarkQueryPerformance1/RoomEnum-4                   1203202       111391        -90.74%
BenchmarkQueryPerformance1/AHUFeed0-4                   1110075       166794        -84.97%
BenchmarkQueryPerformance1/AHUFeed1-4                   1607667       315500        -80.38%
BenchmarkQueryPerformance1/AHUFeed2-4                   1651798       314155        -80.98%
BenchmarkQueryPerformance1/AHUFeed1Reverse-4            1643508       316284        -80.76%
BenchmarkQueryPerformance1/SensorSubclass-4             1435256       125217        -91.28%
BenchmarkQueryPerformance1/VAVExplore-4                 2170243       1379837       -36.42%
```